requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
tldextract
python-whois
flask
tldextract
python-whois
pandas
